Extract follow-button into its own file (IOS-157)
